Feat : article , comment delete 시 user 확인

+++ src/articles/articles.controller.ts
@@ -134,8 +134,28 @@ export class ArticlesController {
   }
 
   @Delete(':id')
+  @ApiResponse({
+    status: 200,
+    description: '삭제 성공',
+  })
+  @ApiResponse({
+    status: 406,
+    description: 'jwt의 유저 정보와 글 작성자가 다른 경우',
+  })
   @UseGuards(AuthGuard('jwt'))
-  remove(@Param('id') id: string, @Request() req) {
+  async remove(@Param('id') id: string, @Request() req) {
+    const article = await this.articlesService.findOne(+id);
+    if (article.writerId !== req.user.id) {
+      throw new HttpException(
+        {
+          status: HttpStatus.NOT_ACCEPTABLE,
+          errors: {
+            msg: 'different user',
+          },
+        },
+        HttpStatus.NOT_ACCEPTABLE,
+      );
+    }
     return this.articlesService.remove(+id);
   }
 }
